Remove need of API key & append lyrics
Pretend to be a webplayer, no need to login or have an api key.
Download lyrics from spotify, this needs a bearer from a logged in webplayer.
The bearer can be found in the browser when pressing f12 (debug) and go to the network tab. Look for a request with a bearer header.
Copy this bearer, and when running the spotifree command append --bearer <bearer>